+++ database/quiz_question_manager.py
@@ -1,22 +1,13 @@
-from sqlalchemy import create_engine, Column, Integer, String, DateTime, Text
+from sqlalchemy import create_engine
 from sqlalchemy.orm import declarative_base
 from sqlalchemy.orm import sessionmaker
 from configuration import sql_file_path
 from datetime import datetime, timezone
 from sqlalchemy.exc import SQLAlchemyError
 from interactions.quiz_question_dto import QuizQuestionDTO
+from models.quiz_question import QuizQuestion
 
 Base = declarative_base()
-
-
-class QuizQuestion(Base):
-    __tablename__ = 'quiz_questions'
-    id = Column(Integer, primary_key=True)
-    question_text = Column(Text)
-    thread_id = Column(String)  # Slack thread ID for tracking conversations
-    summary = Column(Text)  # Summary of the conversation
-    posted_on_slack = Column(DateTime)  # Timestamp when posted on Slack
-    posted_on_confluence = Column(DateTime, nullable=True)  # Timestamp when posted on Confluence
 
 
 class QuizQuestionManager:
