+++ models/base.py
@@ -1,0 +1,13 @@
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy import Column, Integer, String, DateTime, Text
+
+Base = declarative_base()
+
+__all__ = [
+    'Column',
+    'Integer',
+    'String',
+    'DateTime',
+    'Text',
+    'Base'
+]
