+++ models/bookmarked_conversation.py
@@ -1,11 +1,10 @@
-from sqlalchemy import Column, Integer, String, DateTime, Text
-from sqlalchemy.ext.declarative import declarative_base
+from .base import *
 from datetime import datetime, timezone
 
-Base = declarative_base()
 
 class BookmarkedConversation(Base):
     __tablename__ = 'bookmarked_conversations'
+
     id = Column(Integer, primary_key=True)
     title = Column(Text)
     body = Column(Text)
