+++ models/page_data.py
@@ -1,0 +1,21 @@
+from .base import *
+import json
+
+
+class PageData(Base):
+    """
+    SQLAlchemy model for storing Confluence page data.
+    """
+    __tablename__ = 'page_data'
+
+    id = Column(Integer, primary_key=True)
+    page_id = Column(String)
+    space_key = Column(String)
+    title = Column(String)
+    author = Column(String)
+    createdDate = Column(DateTime)
+    lastUpdated = Column(DateTime)
+    content = Column(Text)
+    comments = Column(Text, default=json.dumps([]))
+    last_embedded = Column(DateTime)
+    embed = Column(Text, default=json.dumps([]))
