+++ models/qa_interaction.py
@@ -1,0 +1,19 @@
+from .base import *
+import json
+
+
+class QAInteraction(Base):
+    __tablename__ = 'qa_interactions'
+
+    interaction_id = Column(Integer, primary_key=True)
+    question_text = Column(Text)
+    thread_id = Column(String)
+    assistant_thread_id = Column(String)
+    answer_text = Column(Text)
+    channel_id = Column(String)
+    slack_user_id = Column(String)
+    question_timestamp = Column(DateTime)
+    answer_timestamp = Column(DateTime)
+    comments = Column(Text, default=json.dumps([]))
+    last_embedded = Column(DateTime)
+    embed = Column(Text, default=json.dumps([]))
